backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
attrs==25.4.0
auth==1.3.0
blinker==1.9.0
cbor2==5.8.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
distro==1.9.0
fastapi==0.134.0
Flask==3.1.3
flask-cors==6.0.2
frozenlist==1.8.0
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.72.0
grpclib==0.4.9
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
modal==1.3.4
multidict==6.7.1
oauthlib==3.3.1
openai==1.76.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psycopg==3.3.3
psycopg-binary==3.3.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.3.2
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rsa==4.9.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.47
starlette==0.52.1
synchronicity==0.11.1
toml==0.10.2
tqdm==4.67.3
typer==0.24.1
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
waitress==3.0.2
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.6
yarl==1.22.0
